Client/c_requirements.txt
Pillow
pycryptodome